+++ pyproject.toml
@@ -141,7 +141,7 @@ ignore_missing_imports = true
 [[tool.mypy.overrides]]
 module = [
     "lightly_train._models.dinov2_vit.dinov2_vit_src.*",
-    "lightly_train._models.dinov3_vit.dinov3_vit_src.*",
+    "lightly_train._models.dinov3_vit.dinov3_src.*",
 ]
 ignore_errors = true
