+++ src/lightly_train/_models/dinov3/dinov3_src/hub/utils.py
@@ -1,0 +1,9 @@
+#
+# Copyright (c) Lightly AG and affiliates.
+# All rights reserved.
+#
+# This source code is licensed under the license found in the
+# LICENSE file in the root directory of this source tree.
+#
+
+DINOV3_BASE_URL = "https://dl.fbaipublicfiles.com/dinov3"
