+++ src/lightly_train/_models/dinov3/dinov3_src/layers/__init__.py
@@ -1,0 +1,37 @@
+#
+# Copyright (c) Lightly AG and affiliates.
+# All rights reserved.
+#
+# This source code is licensed under the license found in the
+# LICENSE file in the root directory of this source tree.
+#
+
+# TODO(Lionel, 08/25): Remove the linting skip.
+# ruff: noqa
+
+from lightly_train._models.dinov3.dinov3_src.layers.attention import (
+    CausalSelfAttention,
+    LinearKMaskedBias,
+    SelfAttention,
+)
+from lightly_train._models.dinov3.dinov3_src.layers.block import (
+    CausalSelfAttentionBlock,
+    SelfAttentionBlock,
+)
+from lightly_train._models.dinov3.dinov3_src.layers.ffn_layers import (
+    Mlp,
+    SwiGLUFFN,
+)
+from lightly_train._models.dinov3.dinov3_src.layers.fp8_linear import (
+    convert_linears_to_fp8,
+)
+from lightly_train._models.dinov3.dinov3_src.layers.layer_scale import (
+    LayerScale,
+)
+from lightly_train._models.dinov3.dinov3_src.layers.patch_embed import (
+    PatchEmbed,
+)
+from lightly_train._models.dinov3.dinov3_src.layers.rms_norm import RMSNorm
+from lightly_train._models.dinov3.dinov3_src.layers.rope_position_encoding import (
+    RopePositionEmbedding,
+)
