+++ src/lightly_train/_models/dinov3/dinov3_src/layers/attention.py
@@ -1,8 +1,10 @@
 #
-# Copyright (c) Meta Platforms, Inc. and affiliates.
+# Copyright (c) Lightly AG and affiliates.
+# All rights reserved.
 #
-# This software may be used and distributed in accordance with
-# the terms of the DINOv3 License Agreement.#
+# This source code is licensed under the license found in the
+# LICENSE file in the root directory of this source tree.
+#
 
 import math
 from typing import List, Tuple
@@ -11,7 +13,7 @@ import torch
 import torch.nn.functional as F
 from torch import Tensor, nn
 
-from lightly_train._models.dinov3_vit.dinov3_vit_src.utils import (
+from lightly_train._models.dinov3.dinov3_src.utils import (
     cat_keep_shapes,
     uncat_with_shapes,
 )
