+++ src/lightly_train/_models/dinov3/dinov3_src/layers/fp8_linear.py
@@ -1,17 +1,19 @@
 #
-# Copyright (c) Meta Platforms, Inc. and affiliates.
+# Copyright (c) Lightly AG and affiliates.
+# All rights reserved.
 #
-# This software may be used and distributed in accordance with
-# the terms of the DINOv3 License Agreement.#
+# This source code is licensed under the license found in the
+# LICENSE file in the root directory of this source tree.
+#
 
 import re
 
 import torch
 
-from lightly_train._models.dinov3_vit.dinov3_vit_src.layers.attention import (
+from lightly_train._models.dinov3.dinov3_src.layers.attention import (
     LinearKMaskedBias,
 )
-from lightly_train._models.dinov3_vit.dinov3_vit_src.utils import named_replace
+from lightly_train._models.dinov3.dinov3_src.utils import named_replace
 
 # avoid division by zero when calculating scale
 EPS = 1e-12
