+++ src/lightly_train/_models/dinov3/dinov3_src/layers/sparse_linear.py
@@ -1,8 +1,10 @@
 #
-# Copyright (c) Meta Platforms, Inc. and affiliates.
+# Copyright (c) Lightly AG and affiliates.
+# All rights reserved.
 #
-# This software may be used and distributed in accordance with
-# the terms of the DINOv3 License Agreement.#
+# This source code is licensed under the license found in the
+# LICENSE file in the root directory of this source tree.
+#
 
 import logging
 from typing import Callable
@@ -12,7 +14,7 @@ import torch.nn as nn
 import torch.nn.functional as F
 import xformers.ops as xops
 
-from lightly_train._models.dinov3_vit.dinov3_vit_src.utils import (
+from lightly_train._models.dinov3.dinov3_src.utils import (
     named_apply,
     named_replace,
 )
