+++ src/lightly_train/_models/dinov3/dinov3_src/models/__init__.py
@@ -1,8 +1,10 @@
 #
-# Copyright (c) Meta Platforms, Inc. and affiliates.
+# Copyright (c) Lightly AG and affiliates.
+# All rights reserved.
 #
-# This software may be used and distributed in accordance with
-# the terms of the DINOv3 License Agreement.#
+# This source code is licensed under the license found in the
+# LICENSE file in the root directory of this source tree.
+#
 
 import logging
 from pathlib import Path
@@ -11,10 +13,10 @@ from typing import Union
 import torch
 import torch.nn as nn
 
-from lightly_train._models.dinov3_vit.dinov3_vit_src.layers.fp8_linear import (
+from lightly_train._models.dinov3.dinov3_src.layers.fp8_linear import (
     convert_linears_to_fp8,
 )
-from lightly_train._models.dinov3_vit.dinov3_vit_src.models import (
+from lightly_train._models.dinov3.dinov3_src.models import (
     vision_transformer as vits,
 )
