--- src/lightly_train/_models/dinov3_vit/__init__.py
@@ -1,5 +1,0 @@
-#
-# Copyright (c) Meta Platforms, Inc. and affiliates.
-#
-# This software may be used and distributed in accordance with
-# the terms of the DINOv3 License Agreement.#
