--- src/lightly_train/_models/dinov3_vit/dinov3_vit_src/hub/utils.py
@@ -1,7 +1,0 @@
-#
-# Copyright (c) Meta Platforms, Inc. and affiliates.
-#
-# This software may be used and distributed in accordance with
-# the terms of the DINOv3 License Agreement.#
-
-DINOV3_BASE_URL = "https://dl.fbaipublicfiles.com/dinov3"
