--- src/lightly_train/_models/dinov3_vit/dinov3_vit_src/layers/__init__.py
@@ -1,35 +1,0 @@
-#
-# Copyright (c) Meta Platforms, Inc. and affiliates.
-#
-# This software may be used and distributed in accordance with
-# the terms of the DINOv3 License Agreement.#
-
-# TODO(Lionel, 08/25): Remove the linting skip.
-# ruff: noqa
-
-from lightly_train._models.dinov3_vit.dinov3_vit_src.layers.attention import (
-    CausalSelfAttention,
-    LinearKMaskedBias,
-    SelfAttention,
-)
-from lightly_train._models.dinov3_vit.dinov3_vit_src.layers.block import (
-    CausalSelfAttentionBlock,
-    SelfAttentionBlock,
-)
-from lightly_train._models.dinov3_vit.dinov3_vit_src.layers.ffn_layers import (
-    Mlp,
-    SwiGLUFFN,
-)
-from lightly_train._models.dinov3_vit.dinov3_vit_src.layers.fp8_linear import (
-    convert_linears_to_fp8,
-)
-from lightly_train._models.dinov3_vit.dinov3_vit_src.layers.layer_scale import (
-    LayerScale,
-)
-from lightly_train._models.dinov3_vit.dinov3_vit_src.layers.patch_embed import (
-    PatchEmbed,
-)
-from lightly_train._models.dinov3_vit.dinov3_vit_src.layers.rms_norm import RMSNorm
-from lightly_train._models.dinov3_vit.dinov3_vit_src.layers.rope_position_encoding import (
-    RopePositionEmbedding,
-)
