+++ src/lightly_train/_models/package_helpers.py
@@ -15,7 +15,7 @@ from torch.nn import Module
 from lightly_train._models import model_wrapper
 from lightly_train._models.custom.custom_package import CUSTOM_PACKAGE
 from lightly_train._models.dinov2_vit.dinov2_vit_package import DINOV2_VIT_PACKAGE
-from lightly_train._models.dinov3_vit.dinov3_vit_package import DINOV3_VIT_PACKAGE
+from lightly_train._models.dinov3.dinov3_package import DINOV3_VIT_PACKAGE
 from lightly_train._models.model_wrapper import ModelWrapper
 from lightly_train._models.package import BasePackage, Package
 from lightly_train._models.rfdetr.rfdetr_package import RFDETR_PACKAGE
